// app/src/client/types.gen.ts
export type HTTPValidationError = {
    detail?: Array<ValidationError>;
};

export type OrganizationEmployees = {
    org_id: string;
    organization_name: string;
    organization_description: string;
    data: Array<Employee>;
};

export type TipCreate = {
    employee_id: string;
    amount: number;
};

export type Token = {
    access_token: string;
    token_type?: string;
};

export type ValidationError = {
    loc: Array<(string | number)>;
    msg: string;
    type: string;
};
// ...
export type AddTipApiTipsAddTipPostError = HTTPValidationError;